Move download notification logic into DownloadUtils
